+++ src/game/client/hl2mp/c_hl2mp_player.cpp
@@ -17,7 +17,7 @@
 
 // Don't alias here
 #if defined( CHL2MP_Player )
-#undef CHL2MP_Player	
+#undef CHL2MP_Player
 #endif
 
 LINK_ENTITY_TO_CLASS( player, C_HL2MP_Player );
@@ -82,7 +82,7 @@ void C_HL2MP_Player::UpdateIDTarget()
 	m_iIDEntIndex = 0;
 
 	// don't show IDs in chase spec mode
-	if ( GetObserverMode() == OBS_MODE_CHASE || 
+	if ( GetObserverMode() == OBS_MODE_CHASE ||
 		 GetObserverMode() == OBS_MODE_DEATHCAM )
 		 return;
 
@@ -108,7 +108,7 @@ void C_HL2MP_Player::TraceAttack( const CTakeDamageInfo &info, const Vector &vec
 	Vector vecOrigin = ptr->endpos - vecDir * 4;
 
 	float flDistance = 0.0f;
-	
+
 	if ( info.GetAttacker() )
 	{
 		flDistance = (ptr->endpos - info.GetAttacker()->GetAbsOrigin()).Length();
@@ -119,7 +119,7 @@ void C_HL2MP_Player::TraceAttack( const CTakeDamageInfo &info, const Vector &vec
 		AddMultiDamage( info, this );
 
 		int blood = BloodColor();
-		
+
 		CBaseEntity *pAttacker = info.GetAttacker();
 
 		if ( pAttacker )
@@ -160,7 +160,7 @@ void C_HL2MP_Player::Initialize( void )
 CStudioHdr *C_HL2MP_Player::OnNewModel( void )
 {
 	CStudioHdr *hdr = BaseClass::OnNewModel();
-	
+
 	Initialize( );
 
 	return hdr;
@@ -198,7 +198,7 @@ void C_HL2MP_Player::UpdateLookAt( void )
 
 	float flBodyYawDiff = bodyAngles[YAW] - m_flLastBodyYaw;
 	m_flLastBodyYaw = bodyAngles[YAW];
-	
+
 
 	// Set the head's yaw.
 	float desired = AngleNormalize( desiredAngles[YAW] - bodyAngles[YAW] );
@@ -208,14 +208,14 @@ void C_HL2MP_Player::UpdateLookAt( void )
 	// Counterrotate the head from the body rotation so it doesn't rotate past its target.
 	m_flCurrentHeadYaw = AngleNormalize( m_flCurrentHeadYaw - flBodyYawDiff );
 	desired = clamp( desired, m_headYawMin, m_headYawMax );
-	
+
 	SetPoseParameter( m_headYawPoseParam, m_flCurrentHeadYaw );
 
-	
+
 	// Set the head's yaw.
 	desired = AngleNormalize( desiredAngles[PITCH] );
 	desired = clamp( desired, m_headPitchMin, m_headPitchMax );
-	
+
 	m_flCurrentHeadPitch = ApproachAngle( desired, m_flCurrentHeadPitch, 130 * gpGlobals->frametime );
 	m_flCurrentHeadPitch = AngleNormalize( m_flCurrentHeadPitch );
 	SetPoseParameter( m_headPitchPoseParam, m_flCurrentHeadPitch );
@@ -224,7 +224,7 @@ void C_HL2MP_Player::UpdateLookAt( void )
 void C_HL2MP_Player::ClientThink( void )
 {
 	bool bFoundViewTarget = false;
-	
+
 	Vector vForward;
 	AngleVectors( GetLocalAngles(), &vForward );
 
@@ -241,8 +241,8 @@ void C_HL2MP_Player::ClientThink( void )
 		Vector vMyOrigin =  GetAbsOrigin();
 
 		Vector vDir = vTargetOrigin - vMyOrigin;
-		
-		if ( vDir.Length() > 128 ) 
+
+		if ( vDir.Length() > 128 )
 			continue;
 
 		VectorNormalize( vDir );
@@ -264,7 +264,7 @@ void C_HL2MP_Player::ClientThink( void )
 }
 
 //-----------------------------------------------------------------------------
-// Purpose: 
+// Purpose:
 //-----------------------------------------------------------------------------
 int C_HL2MP_Player::DrawModel( int flags )
 {
@@ -349,7 +349,7 @@ const QAngle &C_HL2MP_Player::EyeAngles()
 }
 
 //-----------------------------------------------------------------------------
-// Purpose: 
+// Purpose:
 //-----------------------------------------------------------------------------
 void C_HL2MP_Player::AddEntity( void )
 {
@@ -359,7 +359,7 @@ void C_HL2MP_Player::AddEntity( void )
 	vTempAngles[PITCH] = m_angEyeAngles[PITCH];
 
 	SetLocalAngles( vTempAngles );
-		
+
 	m_PlayerAnimState.Update();
 
 	// Zero out model pitch, blending takes care of all of it.
@@ -376,12 +376,12 @@ void C_HL2MP_Player::AddEntity( void )
 
 			Vector vecOrigin;
 			QAngle eyeAngles = m_angEyeAngles;
-	
+
 			GetAttachment( iAttachment, vecOrigin, eyeAngles );
 
 			Vector vForward;
 			AngleVectors( eyeAngles, &vForward );
-				
+
 			trace_t tr;
 			UTIL_TraceLine( vecOrigin, vecOrigin + (vForward * 200), MASK_SHOT, this, COLLISION_GROUP_NONE, &tr );
 
@@ -409,7 +409,7 @@ void C_HL2MP_Player::AddEntity( void )
 				beamInfo.m_bRenderable = true;
 				beamInfo.m_flLife = 0.5;
 				beamInfo.m_nFlags = FBEAM_FOREVER | FBEAM_ONLYNOISEONCE | FBEAM_NOTILE | FBEAM_HALOBEAM;
-				
+
 				m_pFlashlightBeam = beams->CreateBeamPoints( beamInfo );
 			}
 
@@ -426,7 +426,7 @@ void C_HL2MP_Player::AddEntity( void )
 
 				dlight_t *el = effects->CL_AllocDlight( 0 );
 				el->origin = tr.endpos;
-				el->radius = 50; 
+				el->radius = 50;
 				el->color.r = 200;
 				el->color.g = 200;
 				el->color.b = 200;
@@ -440,7 +440,7 @@ void C_HL2MP_Player::AddEntity( void )
 	}
 }
 
-ShadowType_t C_HL2MP_Player::ShadowCastType( void ) 
+ShadowType_t C_HL2MP_Player::ShadowCastType( void )
 {
 	if ( !IsVisible() )
 		 return SHADOWS_NONE;
@@ -472,7 +472,7 @@ bool C_HL2MP_Player::ShouldDraw( void )
 
 	if( IsLocalPlayer() && IsRagdoll() )
 		return true;
-	
+
 	if ( IsRagdoll() )
 		return false;
 
@@ -533,8 +533,8 @@ float C_HL2MP_Player::GetFOV( void )
 	float flFOVOffset = C_BasePlayer::GetFOV() + GetZoom();
 
 	// Clamp FOV in MP
-	int min_fov = GetMinFOV();
-	
+	float min_fov = GetMinFOV();
+
 	// Don't let it go too low
 	flFOVOffset = MAX( min_fov, flFOVOffset );
 
@@ -624,7 +624,7 @@ void C_HL2MP_Player::HandleSpeedChanges( void )
 	{
 		if ( IsSuitEquipped() )
 		{
-			// The state of the WALK button has changed. 
+			// The state of the WALK button has changed.
 			if( IsWalking() && !(m_afButtonPressed & IN_WALK) )
 			{
 				StopWalking();
@@ -636,7 +636,7 @@ void C_HL2MP_Player::HandleSpeedChanges( void )
 		}
 	}
 
-	if ( IsSuitEquipped() && m_fIsWalking && !(m_nButtons & IN_WALK)  ) 
+	if ( IsSuitEquipped() && m_fIsWalking && !(m_nButtons & IN_WALK)  )
 		StopWalking();
 }
 
@@ -671,7 +671,7 @@ void C_HL2MP_Player::ItemPreFrame( void )
 	BaseClass::ItemPreFrame();
 
 }
-	
+
 void C_HL2MP_Player::ItemPostFrame( void )
 {
 	if ( GetFlags() & FL_FROZEN )
@@ -691,7 +691,7 @@ void C_HL2MP_Player::CalcView( Vector &eyeOrigin, QAngle &eyeAngles, float &zNea
 {
 	if ( m_lifeState != LIFE_ALIVE && !IsObserver() )
 	{
-		Vector origin = EyePosition();			
+		Vector origin = EyePosition();
 
 		IRagdoll *pRagdoll = GetRepresentativeRagdoll();
 
@@ -704,8 +704,8 @@ void C_HL2MP_Player::CalcView( Vector &eyeOrigin, QAngle &eyeAngles, float &zNea
 		BaseClass::CalcView( eyeOrigin, eyeAngles, zNear, zFar, fov );
 
 		eyeOrigin = origin;
-		
-		Vector vForward; 
+
+		Vector vForward;
 		AngleVectors( eyeAngles, &vForward );
 
 		VectorNormalize( vForward );
@@ -723,7 +723,7 @@ void C_HL2MP_Player::CalcView( Vector &eyeOrigin, QAngle &eyeAngles, float &zNea
 		{
 			eyeOrigin = trace.endpos;
 		}
-		
+
 		return;
 	}
 
@@ -777,10 +777,10 @@ void C_HL2MPRagdoll::Interp_Copy( C_BaseAnimatingOverlay *pSourceEntity )
 {
 	if ( !pSourceEntity )
 		return;
-	
+
 	VarMapping_t *pSrc = pSourceEntity->GetVarMapping();
 	VarMapping_t *pDest = GetVarMapping();
-    	
+
 	// Find all the VarMapEntry_t's that represent the same variable.
 	for ( int i = 0; i < pDest->m_Entries.Count(); i++ )
 	{
@@ -810,21 +810,21 @@ void C_HL2MPRagdoll::ImpactTrace( trace_t *pTrace, int iDamageType, const char *
 	if ( iDamageType == DMG_BLAST )
 	{
 		dir *= 4000;  // adjust impact strenght
-				
+
 		// apply force at object mass center
 		pPhysicsObject->ApplyForceCenter( dir );
 	}
 	else
 	{
-		Vector hitpos;  
-	
+		Vector hitpos;
+
 		VectorMA( pTrace->startpos, pTrace->fraction, dir, hitpos );
 		VectorNormalize( dir );
 
 		dir *= 4000;  // adjust impact strenght
 
 		// apply force where we hit it
-		pPhysicsObject->ApplyForceOffset( dir, hitpos );	
+		pPhysicsObject->ApplyForceOffset( dir, hitpos );
 
 		// Blood spray!
 //		FX_CS_BloodSpray( hitpos, dir, 10 );
@@ -839,7 +839,7 @@ void C_HL2MPRagdoll::CreateHL2MPRagdoll( void )
 	// First, initialize all our data. If we have the player's entity on our client,
 	// then we can make ourselves start out exactly where the player is.
 	C_HL2MP_Player *pPlayer = dynamic_cast< C_HL2MP_Player* >( m_hPlayer.Get() );
-	
+
 	if ( pPlayer && !pPlayer->IsDormant() )
 	{
 		// move my current model instance to the ragdoll's so decals are preserved.
@@ -849,7 +849,7 @@ void C_HL2MPRagdoll::CreateHL2MPRagdoll( void )
 
 		// Copy all the interpolated vars from the player entity.
 		// The entity uses the interpolated history to get bone velocity.
-		bool bRemotePlayer = (pPlayer != C_BasePlayer::GetLocalPlayer());			
+		bool bRemotePlayer = (pPlayer != C_BasePlayer::GetLocalPlayer());
 		if ( bRemotePlayer )
 		{
 			Interp_Copy( pPlayer );
@@ -866,7 +866,7 @@ void C_HL2MPRagdoll::CreateHL2MPRagdoll( void )
 			// This is the local player, so set them in a default
 			// pose and slam their velocity, angles and origin
 			SetAbsOrigin( m_vecRagdollOrigin );
-			
+
 			SetAbsAngles( pPlayer->GetRenderAngles() );
 
 			SetAbsVelocity( m_vecRagdollVelocity );
@@ -877,12 +877,12 @@ void C_HL2MPRagdoll::CreateHL2MPRagdoll( void )
 				Assert( false );	// missing walk_lower?
 				iSeq = 0;
 			}
-			
+
 			SetSequence( iSeq );	// walk_lower, basic pose
 			SetCycle( 0.0 );
 
 			Interp_Reset( varMap );
-		}		
+		}
 	}
 	else
 	{
@@ -894,7 +894,7 @@ void C_HL2MPRagdoll::CreateHL2MPRagdoll( void )
 		SetAbsVelocity( m_vecRagdollVelocity );
 
 		Interp_Reset( GetVarMapping() );
-		
+
 	}
 
 	SetModelIndex( m_nModelIndex );
